renamed test_install.py to test_envfile.py
